:truck: move ConditionalStyledBackground.tsx to components/background
