+++ src/components/common/background/ConditionalStyledBackground.tsx
